Added Feedback if Email is already taken, Added Feedback on insertfail, Added redirect on unknown error

+++ workspaceJSP2/Grasshopper/src/swe2013/dao/UserDAO.java
@@ -17,7 +17,7 @@ public interface UserDAO {
 	User getUserbyEmail(String email);
 	User getUserbyID(long userID);
 
-	long saveUser(User user) throws IllegalArgumentException;
+	int saveUser(User user) throws IllegalArgumentException;
 	
 
 	void deleteUser(User user) throws IllegalArgumentException;
